+++ VideosPage.js
@@ -1,8 +1,7 @@
 import React from "react";
 import { View, Text, FlatList, StyleSheet, ScrollView } from "react-native";
-// import VideoPlayer from "react-native-video";
 import { Get_Videos_Query } from "./queries";
-import Video from "./Video";
+import VideoSingle from "./VideoSingle";
 import { useQuery } from "@apollo/client";
 import client from "./apollo";
 
@@ -23,11 +22,10 @@ const VideosPage = () => {
       );
     }
     const videos = data.allVideos;
-    console.log(videos, "Tits")
     return (
       <ScrollView contentContainerStyle={styles.container}>
         {videos.map((video) => (
-          <Video key={video.id} event={video} />
+          <VideoSingle key={video.id} event={video} />
         ))}
       </ScrollView>
     )
