+++ app.json
@@ -27,4 +27,5 @@
       "favicon": "./assets/favicon.png"
     }
   }
+  
 }
