+++ package.json
@@ -16,6 +16,7 @@
     "@react-navigation/native-stack": "^6.9.13",
     "@react-navigation/stack": "^6.3.17",
     "expo": "~48.0.18",
+    "expo-av": "~13.2.1",
     "expo-image": "~1.0.1",
     "expo-status-bar": "~1.4.4",
     "graphql": "^15.8.0",
@@ -25,7 +26,8 @@
     "react-native-safe-area-context": "4.5.0",
     "react-native-screens": "~3.20.0",
     "react-native-video": "^5.2.1",
-    "react-native-web": "~0.18.10"
+    "react-native-web": "~0.18.10",
+    "react-native-webview": "^13.2.2"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
